# go-swagger-sample/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "`Markdown` Description For Go-Swagger-Sample Api Docs",
        "title": "Title For Go-Swagger-Sample Api Docs",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
        "/users": {
            "get": {
                "description": "查询用户列表描述",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户接口"
                ],
                "summary": "查询用户列表",
                "operationId": "/users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录名称",
                        "name": "username",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/models.Users"
                        }
                    }
                }
            },
            "post": {
                "description": "创建用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户接口"
                ],
                "summary": "创建用户",
                "operationId": "/users",
                "parameters": [
                    {
                        "description": "创建用户请求主体",
                        "name": "createBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "查询用户详情描述",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户接口"
                ],
                "summary": "查询用户详情",
                "operationId": "/users/:id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户标识",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.User": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "用户标识",
                    "type": "string",
                    "example": "1"
                },
                "password": {
                    "description": "登录密码",
                    "type": "string",
                    "example": "123456"
                },
                "username": {
                    "description": "登录名称",
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "models.Users": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "用户列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                },
                "totalcount": {
                    "description": "共计条数",
                    "type": "integer",
                    "example": 0
                }
            }
        }
    }
}